ignore following commands if file doesn't exist

+++ scripts/install/install-qclient.sh
@@ -6,12 +6,14 @@ if [ ! -f "$QUIL_CLIENT_PATH/$CURRENT_QCLIENT_BINARY" ]; then
     qclient_files=$(fetch_available_files "https://releases.quilibrium.com/qclient-release")
     
     get_remote_quil_files qclient_files[@] $QUIL_CLIENT_PATH
-    
-    sudo chmod +x $QUIL_CLIENT_PATH/$CURRENT_QCLIENT_BINARY
 
-    if [ -s $QUIL_QCLIENT_BIN ]; then
-        rm $QUIL_QCLIENT_BIN
+    if [ -f $QUIL_CLIENT_PATH/$CURRENT_QCLIENT_BINARY ]; then
+        sudo chmod +x $QUIL_CLIENT_PATH/$CURRENT_QCLIENT_BINARY
+
+        if [ -s $QUIL_QCLIENT_BIN ]; then
+            rm $QUIL_QCLIENT_BIN
+        fi
+
+        sudo ln -s $QUIL_CLIENT_PATH/$CURRENT_QCLIENT_BINARY $QUIL_QCLIENT_BIN
     fi
-
-    sudo ln -s $QUIL_CLIENT_PATH/$CURRENT_QCLIENT_BINARY $QUIL_QCLIENT_BIN
 fi
